updated names from index.js to functionNames.js across the board for easier error finding
