+++ client/src/pages/employeePage/employeePage.js
@@ -1,0 +1,33 @@
+import React, { useEffect, useState } from "react";
+import Header from "../../components/header/header";
+import Footer from "../../components/footer/footer";
+import ScheduleContainer from "../../components/scheduleContainer/scheduleContainer";
+import Container from "react-bootstrap/Container";
+
+function EmployeePage() {
+  const [schedule, setSchedule] = useState(["temp argument to stop bug"]);
+
+  useEffect(() => {
+    fetch("/api/employee/6041105aad06d732a00f6be4/classes", {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    })
+      .then((res) => res.json())
+      .then((res) => {
+        setSchedule(res);
+      });
+  }, []);
+
+  return (
+    <>
+      <Header />
+      <Container>This Wrapper the two boxes for employees.</Container>
+      <ScheduleContainer />
+      <Footer />
+    </>
+  );
+}
+
+export default EmployeePage;
