+++ client/src/pages/memberPage/memberPage.js
@@ -1,28 +1,28 @@
 import React from "react";
-import Header from "../../components/header";
-import Footer from "../../components/footer";
-import ContentWrapper from "../../components/contentWrapper";
-import MemberInfoBox from "../../components/memberInfoBox";
-import MeetYourTrainerBox from "../../components/meetYourTrainerBox";
+import Header from "../../components/header/header";
+import Footer from "../../components/footer/footer";
+import MemberInfoBox from "../../components/memberInfoBox/memberInfoBox";
+import MeetYourTrainerBox from "../../components/meetYourTrainerBox/meetYourTrainerBox";
 import Row from "react-bootstrap/Row";
 import RenderScheduleHeadings from "../../utilities/render-schedule-headings";
 import AuthenticationButton from "../../components/authenticationButton";
+import Container from "react-bootstrap/Container"
 
 function MemberPage() {
   return (
     <>
       <Header />
       <AuthenticationButton />
-      <ContentWrapper>
+      <Container>
         <Row>
           <MemberInfoBox />
           <MeetYourTrainerBox />
         </Row>
-      </ContentWrapper>
+      </Container>
 
-      <ContentWrapper>
+      <Container>
         <Row>{RenderScheduleHeadings()}</Row>
-      </ContentWrapper>
+      </Container>
 
       <Footer />
     </>
