+++ client/src/pages/registrationPage/registrationPage.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../../components/header";
-import Footer from "../../components/footer";
+import Header from "../../components/header/header";
+import Footer from "../../components/footer/footer";
 import AuthenticationButton from "../../components/authenticationButton";
 
 function RegistrationPage() {
